- hide status bar & navigation bar of splash screen - create onboarding screen & finisg ui - add fonts & imgs with different densities - build selector with two states of onboarding_btn (pressed or not) - add some strings and colors - add lottie dependency to my app - add lottie view in the splash_activity and add text_view
