+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.FoodPlannerApplication"
         tools:targetApi="31">
         <activity
-            android:name=".SplashActivity"
+            android:name=".modules.splash.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,9 +22,8 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".modules.onboarding.OnboardingActivity"
             android:exported="false">
         </activity>
     </application>
-
 </manifest>
