+++ app/src/main/res/drawable/button_background.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="@color/primary"/>
+    <corners android:radius="14dp"/>
+</shape>
