+++ app/src/main/res/values/colors.xml
@@ -3,4 +3,6 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="primary">#FF1CCA69</color>
+    <color name="lighter_primary">#9BEABA</color>
+    <color name="background">#FF0D0D0D</color>
 </resources>
